# Game-Hub/game-Hub-fe/src/openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenApi specification",
    "description": "OpenApi documentation",
    "termsOfService": "Terms of service",
    "contact": {
      "name": "Marcel,",
      "url": "https://google.com",
      "email": "test@gmail.com"
    },
    "license": {
      "name": "License name",
      "url": "https://random-url.com"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8088/api/v1",
      "description": "Local environment"
    },
    {
      "url": "https://api.example.com/v1",
      "description": "Production environment"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/admin/unban/{userId}": {
      "put": {
        "tags": [
          "admin-controller"
        ],
        "operationId": "unBanUser",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          }
        }
      }
    },
    "/admin/reject-report/{reportId}": {
      "put": {
        "tags": [
          "admin-controller"
        ],
        "operationId": "rejectReport",
        "parameters": [
          {
            "name": "reportId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          }
        }
      }
    },
    "/admin/edit-game/{gameId}": {
      "put": {
        "tags": [
          "admin-controller"
        ],
        "operationId": "updateGame",
        "parameters": [
          {
            "name": "gameId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GameUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          }
        }
      }
    },
    "/admin/change-role/{userId}": {
      "put": {
        "tags": [
          "admin-controller"
        ],
        "operationId": "changeRole",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          }
        }
      }
    },
    "/admin/change-report-status/no-action/{reportId}": {
      "put": {
        "tags": [
          "admin-controller"
        ],
        "operationId": "noActionOnReportedUser",
        "parameters": [
          {
            "name": "reportId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          }
        }
      }
    },
    "/admin/change-report-status/in-review/{reportId}": {
      "put": {
        "tags": [
          "admin-controller"
        ],
        "operationId": "changeStatusInReview",
        "parameters": [
          {
            "name": "reportId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          }
        }
      }
    },
    "/admin/ban/{userId}": {
      "put": {
        "tags": [
          "admin-controller"
        ],
        "operationId": "banUser",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BanUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/warn-user/{userId}": {
      "post": {
        "tags": [
          "admin-controller"
        ],
        "operationId": "warnUser",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WarnUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          }
        }
      }
    },
    "/admin/suspend-account/{userId}": {
      "post": {
        "tags": [
          "admin-controller"
        ],
        "operationId": "suspendAccount",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SuspendAccountRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          }
        }
      }
    },
    "/admin/cover/{gameId}": {
      "post": {
        "tags": [
          "admin-controller"
        ],
        "operationId": "uploadGameCoverImage",
        "parameters": [
          {
            "name": "gameId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "file"
                ],
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/admin/add-game": {
      "post": {
        "tags": [
          "admin-controller"
        ],
        "operationId": "addGame",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GameRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/users": {
      "get": {
        "tags": [
          "admin-controller"
        ],
        "operationId": "getAllUsers",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseAdminUserResponse"
                }
              }
            }
          }
        }
      }
    },
    "/admin/user/info/{userId}": {
      "get": {
        "tags": [
          "admin-controller"
        ],
        "operationId": "getUserInfo",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserResponse"
                }
              }
            }
          }
        }
      }
    },
    "/admin/suspended-user/details/{userId}": {
      "get": {
        "tags": [
          "admin-controller"
        ],
        "operationId": "getSuspendedUserDetails",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserModerationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/admin/suspended-accounts": {
      "get": {
        "tags": [
          "admin-controller"
        ],
        "operationId": "getSuspendedAccount",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseAdminSuspendedAccountsResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/reports": {
      "get": {
        "tags": [
          "admin-controller"
        ],
        "operationId": "getAllReports",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseAdminReportsResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/game/info/{gameId}": {
      "get": {
        "tags": [
          "admin-controller"
        ],
        "operationId": "getGameInfo",
        "parameters": [
          {
            "name": "gameId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameResponse"
                }
              }
            }
          }
        }
      }
    },
    "/admin/dashboard": {
      "get": {
        "tags": [
          "admin-controller"
        ],
        "operationId": "loadDashboardData",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardResponse"
                }
              }
            }
          }
        }
      }
    },
    "/admin/admin/games": {
      "get": {
        "tags": [
          "admin-controller"
        ],
        "operationId": "getAllGames",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseGamePreviewResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/game/delete/{gameId}": {
      "delete": {
        "tags": [
          "admin-controller"
        ],
        "operationId": "deleteGame",
        "parameters": [
          {
            "name": "gameId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "GameUpdateRequest": {
        "required": [
          "ageRatingId",
          "cpu",
          "description",
          "developer",
          "gameUnitSize",
          "genresIds",
          "gpu",
          "languageIds",
          "platformIds",
          "price",
          "publisher",
          "ram",
          "releaseYear",
          "storage",
          "subtitleIds",
          "title"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "publisher": {
            "type": "string"
          },
          "developer": {
            "type": "string"
          },
          "releaseYear": {
            "type": "string"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "discountPercent": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "cpu": {
            "type": "string"
          },
          "gpu": {
            "type": "string"
          },
          "ram": {
            "type": "integer",
            "format": "int32"
          },
          "storage": {
            "type": "integer",
            "format": "int32"
          },
          "gameUnitSize": {
            "type": "string"
          },
          "platformIds": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          },
          "languageIds": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          },
          "subtitleIds": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          },
          "genresIds": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          },
          "ageRatingId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "BanUserRequest": {
        "required": [
          "banReason"
        ],
        "type": "object",
        "properties": {
          "banReason": {
            "type": "integer",
            "format": "int64"
          },
          "reportId": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "customMessage": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "WarnUserRequest": {
        "required": [
          "customMsg",
          "reportId"
        ],
        "type": "object",
        "properties": {
          "customMsg": {
            "type": "string"
          },
          "reportId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "SuspendAccountRequest": {
        "required": [
          "customMessage",
          "expiresAt",
          "reportId",
          "suspendReason"
        ],
        "type": "object",
        "properties": {
          "suspendReason": {
            "type": "integer",
            "format": "int64"
          },
          "customMessage": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string"
          },
          "reportId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "GameRequest": {
        "required": [
          "ageRatingId",
          "cpu",
          "description",
          "developer",
          "gameUnitSize",
          "genresIds",
          "gpu",
          "languageIds",
          "platformIds",
          "price",
          "publisher",
          "ram",
          "releaseYear",
          "storage",
          "subtitleIds",
          "title"
        ],
        "type": "object",
        "properties": {
          "gameId": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "genresIds": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          },
          "description": {
            "type": "string"
          },
          "publisher": {
            "type": "string"
          },
          "developer": {
            "type": "string"
          },
          "releaseYear": {
            "type": "string"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "cpu": {
            "type": "string"
          },
          "gpu": {
            "type": "string"
          },
          "ram": {
            "type": "integer",
            "format": "int32"
          },
          "storage": {
            "type": "integer",
            "format": "int32"
          },
          "gameUnitSize": {
            "type": "string"
          },
          "platformIds": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          },
          "languageIds": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          },
          "subtitleIds": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          },
          "ageRatingId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "LocationResponse": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "iconPath": {
            "type": "string"
          }
        }
      },
...
      "AdminUserResponse": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "userName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "profilePicture": {
            "type": "string",
            "format": "byte"
          },
          "profileColor": {
            "type": "string"
          },
          "bio": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "USER",
              "ADMIN"
            ]
          },
          "accountType": {
            "type": "string",
            "enum": [
              "ADULT",
              "CHILD"
            ]
          },
          "accountLevel": {
            "type": "integer",
            "format": "int64"
          },
          "location": {
            "$ref": "#/components/schemas/LocationResponse"
          },
          "accountStatus": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "BANNED",
              "DISABLED",
              "SUSPENDED",
              "DELETING",
              "DELETED"
            ]
          },
          "registered": {
            "type": "string",
            "format": "date-time"
          },
          "lastLogin": {
            "type": "string",
            "format": "date-time"
          },
          "lastModifiedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "PageResponseAdminUserResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminUserResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          }
        }
      },
...
      "AdminUserModerationResponse": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "userName": {
            "type": "string"
          },
          "accountStatus": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "BANNED",
              "DISABLED",
              "SUSPENDED",
              "DELETING",
              "DELETED"
            ]
          },
          "registeredAt": {
            "type": "string",
            "format": "date-time"
          },
          "reportCount": {
            "type": "integer",
            "format": "int64"
          },
          "warningCount": {
            "type": "integer",
            "format": "int64"
          },
          "suspensionCount": {
            "type": "integer",
            "format": "int64"
          },
          "banCount": {
            "type": "integer",
            "format": "int64"
          },
          "reports": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminReportsResponse"
            }
          },
          "warnings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminWarningResponse"
            }
          },
          "suspensions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminSuspendedAccountsResponse"
            }
          },
          "bans": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminBanResponse"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT auth",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}